+++ index.php
@@ -5,12 +5,14 @@ define('ROOT', __DIR__);
 use App\Controllers\NewsController;
 use App\Controllers\CommentController;
 use App\Repositories\NewsRepository;
+use App\Repositories\CommentRepository;
 
+$commentRepository = new CommentRepository();
 
-foreach (NewsController::getInstance(new NewsRepository())->listNews() as $news) {
+foreach (NewsController::getInstance(new NewsRepository(), $commentRepository)->listNews() as $news) {
 	echo ("############ NEWS " . $news->getTitle() . " ############\n");
 	echo ($news->getBody() . "\n");
-	foreach (CommentController::getInstance()->listComments() as $comment) {
+	foreach (CommentController::getInstance($commentRepository)->listComments() as $comment) {
 		if ($comment->getNewsId() == $news->getId()) {
 			echo ("Comment " . $comment->getId() . " : " . $comment->getBody() . "\n");
 		}
@@ -18,5 +20,5 @@ foreach (NewsController::getInstance(new NewsRepository())->listNews() as $news)
 }
 
 
-$commentManager = CommentController::getInstance();
-$c = $commentManager->listComments();
+// $commentManager = CommentController::getInstance();
+// $c = $commentManager->listComments();